requirements.txt
matplotlib
pywinauto
pyserial
saleae